Revert "Add files via upload"
This reverts commit c7c841aaaaee169515dab30851e87e727b0ac44c.

+++ A6_PVTranslatorCloud/src/viewsCampanya.py
@@ -29,13 +29,13 @@ class DeleteCampanya(BaseHandler):
     
     def get(self,id_campanya):
        
-        idCampanya = int(id_campanya)
+        if users.is_current_user_admin():
+            idCampanya = int(id_campanya)
+            campanya = db.get(db.Key.from_path('Campanyas', idCampanya))
+            id_mod = str(campanya.modulo)
+            db.delete(campanya)
         
-        campanya = db.get(db.Key.from_path('Campanyas', idCampanya))
-        id_mod = str(campanya.modulo)
-        db.delete(campanya)
         time.sleep(0.1)
-        
         return webapp2.redirect('/campanyas/'+id_mod)
        
         
@@ -44,7 +44,10 @@ class DeleteCampanya(BaseHandler):
 class NewCampaign(BaseHandler):
 
     def get(self,id_modulo):
-        self.render_template('newCampanya.html', {"id_modulo" : id_modulo})
+        if users.is_current_user_admin():
+            self.render_template('newCampanya.html', {"id_modulo" : id_modulo})
+        else:
+            return webapp2.redirect('/campanyas/'+id_modulo)
         
     def post(self,id_modulo):
         campanya = Campanyas(name=self.request.get('inputName'),modulo=int(id_modulo),)
@@ -59,8 +62,11 @@ class EditCampaign(BaseHandler):
     def get(self, camp_id):
         id = int(camp_id)
         campanya = db.get(db.Key.from_path('Campanyas', id))
-        self.render_template('newCampanya.html', {"campanya": campanya,"id_modulo" : campanya.modulo})
-        
+            
+        if users.is_current_user_admin():
+            self.render_template('newCampanya.html', {"campanya": campanya,"id_modulo" : campanya.modulo})
+        else:
+            return webapp2.redirect('/campanyas/'+campanya.modulo)
         
     def post(self, camp_id):
         campanya = db.get(db.Key.from_path('Campanyas', int(camp_id)))
